==> 0.outcomes/0.New_analysis_cases_with_OA/6.statistical-test_treatmentANDcontrol.ipynb <==
import numpy as np
import pandas as pd
from scipy.stats import chi2_contingency
import matplotlib.pyplot as plt
from scipy.stats import mannwhitneyu
import pandas as pd
from statsmodels.stats.proportion import proportions_ztest

import matplotlib.pyplot as plt
import seaborn as sns

#### Physical activity

dfPA_treatment = pd.read_csv("testdata/2.PA_treatment.csv")
dfPA_control = pd.read_csv("testdata/2.PA_control.csv")
dfPA_balanced = pd.read_csv("testdata/3.PA_balanced.csv")

# List of file paths and corresponding dataframe names
files = [
    ("testdata/2.PA_treatment.csv", "dfPA_treatment"),
    ("testdata/2.PA_control.csv", "dfPA_control"),
    ("testdata/3.PA_balanced.csv", "dfPA_balanced"),
]

# Dictionary to store dataframes
dataframes = {}

# Loop through each file and read the csv into a dataframe, then store it in the dictionary
for file, name in files:
    dataframes[name] = pd.read_csv(file)
    print(f"{name}: {dataframes[name].shape}")

##### Create column outcome in categorical:

# Define the function to categorize PA_changes
def categorize_change(PA_change):
    if PA_change > 0:
        return 'Increased'
    elif PA_change < 0:
        return 'Decreased'
    else:
        return 'Unchanged'

# Apply the function to create a new column in df_PA_control_before
dfPA_treatment['PA_change_category'] = dfPA_treatment['PA_change'].apply(categorize_change)
dfPA_control['PA_change_category'] = dfPA_control['PA_change'].apply(categorize_change)

##### Calculate Chi2:

dfPA_treatment['Group'] = 'Treatment'
dfPA_control['Group'] = 'Control'


df_combined = pd.concat([dfPA_treatment, dfPA_control])

# Create a contingency table for PA_change_category across Treatment and Control groups
contingency_table = pd.crosstab(df_combined['PA_change_category'], df_combined['Group'])

# Perform chi-squared test
chi2, p, dof, ex = chi2_contingency(contingency_table)

print(f"Chi-squared Test for PA_change_category across Treatment and Control groups")
print(f"Chi2: {chi2}, p-value: {p}, Degrees of Freedom: {dof}")
print("\nExpected Frequencies:")
print(ex)
print("\nObserved Frequencies:")
print(contingency_table)


#### Quality of life

dfQ_treatment = pd.read_csv("testdata/2.QOL_treatment.csv")
dfQ_control = pd.read_csv("testdata/2.QOL_control.csv")
dfQ_balanced = pd.read_csv("testdata/3.QOL_balanced.csv")

# List of file paths and corresponding dataframe names
files = [
    ("testdata/2.QOL_treatment.csv", "dfQ_treatment"),
    ("testdata/2.QOL_control.csv", "dfQ_control"),
    ("testdata/3.QOL_balanced.csv", "dfQ_balanced"),
]

# Dictionary to store dataframes
dataframes = {}

# Loop through each file and read the csv into a dataframe, then store it in the dictionary
for file, name in files:
    dataframes[name] = pd.read_csv(file)
    print(f"{name}: {dataframes[name].shape}")

# Define the function to categorize PA_changes
def QOL_categorize_significant_change(QOL_change):
    if QOL_change >= 10:
        return 'Increased 10+ points'
    elif QOL_change <= -10:
        return 'Decreased 10+ points'
    else:
        return 'Maintained'

dfQ_treatment['QOL_change_category'] = dfQ_treatment['QOL_change'].apply(QOL_categorize_significant_change)
dfQ_control['QOL_change_category'] = dfQ_control['QOL_change'].apply(QOL_categorize_significant_change)
dfQ_balanced['QOL_change_category'] = dfQ_balanced['QOL_change'].apply(QOL_categorize_significant_change)

dfQ_control

# Add a new column to identify the group
dfQ_treatment['Group'] = 'Treatment'
dfQ_control['Group'] = 'Control'

# Combine the dataframes
dfQ_combined = pd.concat([dfQ_treatment, dfQ_control])

# Create a contingency table for QOL_change_category across Treatment and Control groups
contingency_table = pd.crosstab(dfQ_combined['QOL_change_category'], dfQ_combined['Group'])

# Perform chi-squared test
chi2, p, dof, ex = chi2_contingency(contingency_table)

# Display results
print(f"Chi-squared Test for QOL_change_category across Treatment and Control groups")
print(f"Chi2: {chi2}, p-value: {p}, Degrees of Freedom: {dof}")
print("\nExpected Frequencies:")
print(ex)
print("\nObserved Frequencies:")
print(contingency_table)


#### Pain

dfpain_treatment = pd.read_csv("testdata/2.Pain_treatment.csv")
dfpain_control = pd.read_csv("testdata/2.Pain_control.csv")
dfpain_balanced = pd.read_csv("testdata/3.Pain_balanced.csv")

# List of file paths and corresponding dataframe names
files = [
    ("testdata/2.Pain_treatment.csv", "dfpain_treatment"),
    ("testdata/2.Pain_control.csv", "dfpain_control"),
    ("testdata/3.Pain_balanced.csv", "dfpain_balanced")
]

# Dictionary to store dataframes
dataframes = {}

# Loop through each file and read the csv into a dataframe, then store it in the dictionary
for file, name in files:
    dataframes[name] = pd.read_csv(file)
    print(f"{name}: {dataframes[name].shape}")


# Define the function to categorize changes of at least 20 points
def pain_categorize_significant_change(pain_change):
    if pain_change <= -20:
        return 'Decreased 20+ points'
    elif pain_change >= 20:
        return 'Increased 20+ points'
    else:
        return 'Maintained'  # This now clearly means the pain changed but less than 20 points either way

dfpain_treatment['pain_change_category'] = dfpain_treatment['pain_change'].apply(pain_categorize_significant_change)
dfpain_control['pain_change_category'] = dfpain_control['pain_change'].apply(pain_categorize_significant_change)
dfpain_balanced['pain_change_category'] = dfpain_balanced['pain_change'].apply(pain_categorize_significant_change)


# Add a new column to identify the group
dfpain_treatment['Group'] = 'Treatment'
dfpain_control['Group'] = 'Control'

# Combine the dataframes
dfpain_combined = pd.concat([dfpain_treatment, dfpain_control])

# Create a contingency table for pain_change_category across Treatment and Control groups
contingency_table = pd.crosstab(dfpain_combined['pain_change_category'], dfpain_combined['Group'])

# Perform chi-squared test
chi2, p, dof, ex = chi2_contingency(contingency_table)

# Display results
print(f"Chi-squared Test for pain_change_category across Treatment and Control groups")
print(f"Chi2: {chi2}, p-value: {p}, Degrees of Freedom: {dof}")
print("\nExpected Frequencies:")
print(ex)
print("\nObserved Frequencies:")
print(contingency_table)
