==> 0.outcomes/0.New_analysis_cases_with_OA/8.merge_oai-for new variables discussion.ipynb <==
import pandas as pd
from functools import reduce

df1 = pd.read_sas('testdata/enrollees.sas7bdat')
df2 = pd.read_sas('testdata/measinventory.sas7bdat')
df3 = pd.read_sas('testdata/outcomes99.sas7bdat')
df4 = pd.read_sas('testdata/subjectchar00.sas7bdat')
df5 = pd.read_sas('testdata/allclinical00.sas7bdat')
df6 = pd.read_sas('testdata/allclinical01.sas7bdat')
df7 = pd.read_sas('testdata/allclinical10.sas7bdat')
print("OAI df1 shape:", df1.shape)
print("OAI df2 shape:", df2.shape)
print("OAI df3 shape:", df3.shape)
print("OAI df4 shape:", df4.shape)
print("OAI df5 shape:", df5.shape)
print("OAI df6 shape:", df6.shape)
print("OAI df7 shape:", df7.shape)

# uppercase all of variables
dfs = [df1, df2, df3, df4, df5, df6, df7]

for df in dfs:
    df.columns = df.columns.str.upper()

##### Variables finding

variables = [
    "V00AGE", 
    "P02SEX", 
    "P01BMI", 
    "V00CESD6", 
    "V00P7LKRCV", 
    "V00P7RKRCV", 
    "V00KOOSQOL", 
    "V00CEMPLOY", 
    "V00PASE", 
    "V00COHORT",
    "V01P7LKRCV",
    "V01P7RKRCV",
    "V01KOOSQOL",
    "V01PASE",
    
    'V00CHIRCUR', 'V10CHIRCUR', #baseline & followup Seeing chiropractic care practitioner 
    'V00HOMECV', 'V10HOMECV' ,  #baseline & followup Seen homeopathy/homeopathic treatment practitioner
    
    'V00NSAIDS','V10NSAIDS',     #baseline & followup medication nonprescription
    
    'V00NSAIDRX', 'V10NSAIDRX',  #baseline & followup medication prescription
    
    
#Surgery baseline
"P02KSURG",
"P01KSURGR",
"P01KRSR",
"P01ARTR",
"P01OTSURGR",
"P01KSURGL",
"P01ARTL",
#Surgery followup
"V01KSRGR12",
"V01SREPRKR",
"V01KRSR12",
"V01ARTR12",
"V01KSRGL12",
"V01KRSL12"
    
#Replacement surgery
,"V00EKRSL" ,"V01EKRSL" ,"V00EKRSR" ,"V01EKRSR"  
    
]

result = {}

for i, df in enumerate(dfs, start=1):
    df_name = f"df{i}"
    present_vars = [var for var in variables if var in df.columns]
    result[df_name] = present_vars


result

variable_presence = {var: False for var in variables}

for var in variables:
    for df in dfs:
        if var in df.columns:
            variable_presence[var] = True
            break  


variable_presence

##### We understand we need df1 and df5 and df6 (followup variables) which is enrollees and allclinical00

df1 = df1.drop(columns=['VERSION'])
df5 = df5.drop(columns=['VERSION'])
df6 = df6.drop(columns=['VERSION'])
df7 = df7.drop(columns=['VERSION'])
print("OAI df1 shape:", df1.shape)
print("OAI df5 shape:", df5.shape)
print("OAI df6 shape:", df6.shape)
print("OAI df7 shape:", df7.shape)

dfx = df1[["ID", "P02SEX","V00COHORT"]]
dfx

dfy = df5[["ID", "V00AGE", "P01BMI", "V00CESD6", "V00P7LKRCV", "V00P7RKRCV", "V00KOOSQOL", "V00CEMPLOY", "V00PASE", 
        
        'V00CHIRCUR','V00HOMECV', 'V00NSAIDS','V00NSAIDRX',
        
        "P02KSURG", "P01KSURGR","P01KRSR","P01ARTR","P01OTSURGR","P01KSURGL","P01ARTL",
        
        'V00EKRSL', 'V00EKRSR'
    
        ]]
 
dfy

dfz = df6[['ID', 'V01P7LKRCV', 'V01P7RKRCV', 'V01KOOSQOL', 'V01PASE',
           
           "V01KSRGR12","V01SREPRKR","V01KRSR12","V01ARTR12","V01KSRGL12","V01KRSL12",
           
           'V01EKRSL', 'V01EKRSR'
           ]]
dfz

dfQ = df7[['ID','V10CHIRCUR', 'V10HOMECV', 'V10NSAIDS', 'V10NSAIDRX']]

dfz = df6[['ID', 'V01P7LKRCV', 'V01P7RKRCV', 'V01KOOSQOL', 'V01PASE',
           
           "V01KSRGR12","V01SREPRKR","V01KRSR12","V01ARTR12","V01KSRGL12","V01KRSL12",
           
           'V01EKRSL', 'V01EKRSR'
           ]]
dfz

#Common columns between the dataframes
common_columns = set(dfx.columns) & set(dfy.columns) & set(dfz.columns)& set(dfQ.columns)

common_columns_list = list(common_columns)
common_columns

# Merge 3 datasets:
dfs = [dfx, dfy, dfz, dfQ]
df = reduce(lambda left, right: pd.merge(left, right, on='ID'), dfs)
print("OAI shape after merging 2 datasets:", df.shape)
df

df.V01EKRSL

df.to_csv('testdata/0.OAI18Feb.csv')
# df.to_csv('testdata/0.OAI.csv')

df = pd.read_csv('testdata/0.OAI.csv')
